feat: add support for loading a golang wasm module
