+++ src/App/App.tsx
@@ -10,6 +10,7 @@ import { YamlEditor } from './YamlEditor';
 interface Props {}
 
 export function App(props: Props) {
+  console.log('loaded app');
   const classes = useClasses(props);
   const dispatch = useAppDispatch();
   const { yamlSchema, yamlInput } = useAppSelector(selectAppSlice);
