+++ src/LoadWasm/types.d.ts
@@ -1,0 +1,8 @@
+declare global {
+  export interface Window {
+    Go: any;
+    getHtml: () => string;
+  }
+}
+
+export {};
